add styled component
