+++ api/routes/validationChain.js
@@ -1,4 +1,4 @@
-const { check, validationResult } = require('express-validator');
+const { check } = require('express-validator');
 const { User } = require('../models');
 // validation for route
 const validationChain = [
@@ -28,28 +28,4 @@ const validationChain = [
     .isLength({ min: 8 })
     .withMessage('Please provide a password that is at least 8 chars long'),
 ];
-const valdationMiddleWare = async (req, res, next) => {
-  const errors = validationResult(req);
-
-  if (!errors.isEmpty()) {
-    const errorMessages = errors.array().map(err => err.msg);
-    res.status(400).json({ errors: errorMessages });
-  } else {
-    next();
-  }
-};
-
-const courseValidationChain = [
-  check('title')
-    .exists({ checkNull: true, checkFalsy: true })
-    .withMessage('Please provide a Course Title'),
-  check('description')
-    .exists({ checkNull: true, checkFalsy: true })
-    .withMessage('Please provide a Course Description'),
-  valdationMiddleWare,
-];
-
-module.exports = {
-  validationChain,
-  courseValidationChain,
-};
+exports.validationChain = validationChain;
